tools: update remark configuration
Specify the plugin name as 'remark-lint/', as new remark-cli refused to
find it when cwd is different from the directory where node_modules are
put. Trailing slash fixes this, as the plugin name gets treated as a
path, so `require` works on it.

Explicitly specify the list of rules we want to enable, as the new
remark-lint is opt-in and doesn't come with any rules by default.

Reorder the rules alphabetically.

PR-URL: https://github.com/nodejs/node/pull/8666
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Rich Trott <rtrott@gmail.com>